correccion 2.3.6.9 de edicion luces

+++ backend/index.js
@@ -1249,7 +1249,7 @@ app.post("/api/editar-luz", async (req, res) => {
         proveedor || "",
         parseFloat(costo),
         parseFloat(precio),
-        parseInt(stock),
+        parseFloat(stock),
         id,
       ]
     );
